// src/app/Features/users-list/Interfaces/user.interface.ts
export interface User{
  id: number;
  userName: string;
  email: string;
  country: string;
  phoneNumber: string;
  imagePath: null;
  group: Group;
  creationDate: string;
  modificationDate: string;}

interface Group {
  id: number;
  name: string;
  creationDate: string;
  modificationDate: string;

}